Refactor util to resource package
