+++ cli/commands/create/create.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/sensu/sensu-go/cli"
-	"github.com/sensu/sensu-go/cli/util"
+	"github.com/sensu/sensu-go/cli/resource"
 	"github.com/spf13/cobra"
 )
 
@@ -36,16 +36,16 @@ func execute(cli *cli.SensuCli) func(*cobra.Command, []string) error {
 		if err != nil {
 			return err
 		}
-		processor := util.NewPutter()
+		processor := resource.NewPutter()
 		if len(inputs) == 0 {
-			return util.ProcessStdin(cli, client, processor)
+			return resource.ProcessStdin(cli, client, processor)
 		}
 		recurse, err := cmd.Flags().GetBool("recursive")
 		if err != nil {
 			return err
 		}
 		for _, input := range inputs {
-			if err := util.Process(cli, client, input, recurse, processor); err != nil {
+			if err := resource.Process(cli, client, input, recurse, processor); err != nil {
 				return err
 			}
 		}
