+++ cli/resource/process_test.go
@@ -1,4 +1,4 @@
-package util
+package resource
 
 import (
 	"bytes"
@@ -13,7 +13,7 @@ import (
 	"github.com/sensu/sensu-go/types"
 )
 
-func TestValidateResources(t *testing.T) {
+func TestValidate(t *testing.T) {
 	tests := []struct {
 		name          string
 		resource      *types.Wrapper
@@ -57,19 +57,19 @@ func TestValidateResources(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			resources := []*types.Wrapper{tt.resource}
-			_ = ValidateResources(resources, tt.namespace)
+			_ = Validate(resources, tt.namespace)
 
 			if tt.resource.ObjectMeta.Namespace != tt.wantNamespace {
-				t.Errorf("ValidateResources() wrapper namespace = %q, want namespace %q", tt.resource.ObjectMeta.Namespace, tt.wantNamespace)
+				t.Errorf("Validate() wrapper namespace = %q, want namespace %q", tt.resource.ObjectMeta.Namespace, tt.wantNamespace)
 			}
 			if tt.resource.Value != nil && tt.resource.Value.GetObjectMeta().Namespace != tt.wantNamespace {
-				t.Errorf("ValidateResources() wrapper's resource namespace = %q, want namespace %q", tt.resource.Value.GetObjectMeta().Namespace, tt.wantNamespace)
+				t.Errorf("Validate() wrapper's resource namespace = %q, want namespace %q", tt.resource.Value.GetObjectMeta().Namespace, tt.wantNamespace)
 			}
 		})
 	}
 }
 
-func TestValidateResourcesStderr(t *testing.T) {
+func TestValidateStderr(t *testing.T) {
 	oldStderr := os.Stderr
 	r, w, _ := os.Pipe()
 	os.Stderr = w
@@ -85,7 +85,7 @@ func TestValidateResourcesStderr(t *testing.T) {
 	resources := []*types.Wrapper{&types.Wrapper{
 		ObjectMeta: corev2.NewObjectMeta("check-cpu", "default"),
 	}}
-	_ = ValidateResources(resources, "default")
+	_ = Validate(resources, "default")
 
 	// Reset stderr
 	w.Close()
@@ -95,7 +95,7 @@ func TestValidateResourcesStderr(t *testing.T) {
 	errMsg = strings.TrimSpace(errMsg)
 	wantErr := `error validating resource #0 with name "check-cpu" and namespace "default": resource is nil`
 	if errMsg != wantErr {
-		t.Errorf("ValidateResources() err = %s, want %s", errMsg, wantErr)
+		t.Errorf("Validate() err = %s, want %s", errMsg, wantErr)
 	}
 }
 
@@ -150,16 +150,16 @@ func TestCreatedByLabel(t *testing.T) {
 			bytes, _ := json.Marshal(&tt.resource)
 			stdin.Write(bytes)
 
-			got, err := ParseResources(&stdin)
+			got, err := Parse(&stdin)
 			if err != nil {
-				t.Errorf("ParseResources() error = %v", err)
+				t.Errorf("Parse() error = %v", err)
 				return
 			}
 			if !reflect.DeepEqual(got[0].ObjectMeta.Labels, tt.want) {
-				t.Errorf("ParseResources() inner labels = %v, want %v", got[0].ObjectMeta.Labels, tt.want)
+				t.Errorf("Parse() inner labels = %v, want %v", got[0].ObjectMeta.Labels, tt.want)
 			}
 			if !reflect.DeepEqual(got[0].Value.GetObjectMeta().Labels, tt.want) {
-				t.Errorf("ParseResources() outer labels = %v, want %v", got[0].Value.GetObjectMeta().Labels, tt.want)
+				t.Errorf("Parse() outer labels = %v, want %v", got[0].Value.GetObjectMeta().Labels, tt.want)
 			}
 		})
 	}
